+++ packages/react-native-editor/__device-tests__/pages/editor-page.js
@@ -570,6 +570,17 @@ class EditorPage {
 		return await typeString( this.driver, textViewElement, text, clear );
 	}
 
+	async closePicker() {
+		if ( isAndroid() ) {
+			await swipeDown( this.driver );
+		} else {
+			const cancelButton = await this.driver.elementByAccessibilityId(
+				'Cancel'
+			);
+			await cancelButton.click();
+		}
+	}
+
 	// =============================
 	// Unsupported Block functions
 	// =============================