+++ packages/react-native-editor/CHANGELOG.md
@@ -13,6 +13,10 @@ For each user feature we should also add a importance categorization label  to i
 
 -   [*] Image block: "Set as featured" button within image block settings. (Android and dev only) [#28854]
 -   [*] Image block: Add a "featured" banner. (Android only) [#30806]
+-   [**] The media upload options of the Image, Video and Gallery block automatically opens when the respective block is inserted. [#29546]
+-   [**] The media upload options of the File and Audio block automatically opens when the respective block is inserted. [#31025]
+
+
 
 ## 1.51.0